another validation attempt

+++ express-api/src/controllers/properties/propertiesController.ts
@@ -192,8 +192,11 @@ export const importProperties = async (req: Request, res: Response) => {
   try {
     readFile(filePath, { WTF: true }); //With this read option disabled it will throw if unexpected data is present.
   } catch (e) {
-    const realPath = fs.realpathSync(path.resolve(__dirname + '/../../../', filePath));
-    fs.unlinkSync(realPath);
+    const rootPath = path.resolve(__dirname + '/../../../uploads');
+    const realPath = fs.realpathSync(path.resolve(filePath));
+    if (realPath.startsWith(rootPath)) {
+      fs.unlinkSync(realPath);
+    }
     return res.status(400).send(e.message);
   }
